+++ services/common.js
@@ -115,15 +115,11 @@ export const getAPI = async (path, locale) => {
     return response ? response.data : null
 }
 
-export const getMockAPI = async (path, locale) => {
-    const nextPrefix = path.indexOf('?') > -1 ? '&' : '?'
-    const locQs = locale
-        ? `${nextPrefix}_locale=${locale}`
-        : `${nextPrefix}_locale=${defaultLang}`
+export const getMockAPI = async (path) => {
     const response = await axios
-        .get(`https://mock-api.haloida.dev/v1/${path}${locQs}`)
+        .get(`https://mock-api.haloida.dev/v1/${path}`)
         .catch(function (error) {
-            console.error(`Error at ${path} - ${error}`)
+            console.error(`Error at mock ${path} - ${error}`)
         })
 
     return response ? response.data : null